Sort of builds for platformIO

+++ CPlusPlus/lib/IIntegerConsumer.h
@@ -5,7 +5,7 @@
 
 class IIntegerConsumer {
     public:
-    virtual void Consume(int32_t value) = 0;
+    virtual void Consume(int16_t value) = 0;
 };
 
 #endif /* C5B35687_247F_418A_966A_7B1BC46C1D03 */
